rebased DataPackage on new Resource
